feat(idpe-17265): refactor to check env var only once and set the appropriate httpDelegate handler, and to divide the v1 versus v2 route handling into separate modules.
